+++ paddle/cinn/hlir/dialect/operator/ir/generate_shape_util.h
@@ -19,7 +19,7 @@
 #include <vector>
 #include "paddle/cinn/hlir/dialect/operator/ir/manual_op.h"
 #include "paddle/pir/core/builder.h"
-#include "paddle/pir/dialect/shape/utils/dim_expr.h"
+#include "paddle/pir/dialect/shape/utils/shape_or_data_expr.h"
 
 namespace cinn::dialect {
