+++ paddle/pir/dialect/shape/ir/shape_attribute_storage.h
@@ -23,7 +23,7 @@
 #include "paddle/pir/core/attribute_base.h"
 #include "paddle/pir/core/type.h"
 #include "paddle/pir/core/utils.h"
-#include "paddle/pir/dialect/shape/utils/dim_expr.h"
+#include "paddle/pir/dialect/shape/utils/shape_or_data_expr.h"
 
 namespace pir::shape {
